anomaly-worker/app/storage.py
import psycopg2
from psycopg2.extras import RealDictCursor


class Database:
    """PostgreSQL database interface"""

    def __init__(self, connection_string):
        self.connection_string = connection_string
        self.conn = None

    def connect(self):
        """Establish database connection"""
        self.conn = psycopg2.connect(self.connection_string)

    def create_incident(self, target_id, kind, details):
        """Create a new incident record"""
        # TODO: Implement incident creation
        pass

    def get_recent_probes(self, target_id, limit=100):
        """Get recent probe results for a target"""
        # TODO: Implement probe history retrieval
        pass

    def close(self):
        """Close database connection"""
        if self.conn:
            self.conn.close()
