anomaly-worker/requirements.txt
asyncio-nats-client==0.11.5
psycopg2-binary==2.9.9
numpy==1.26.2
scipy==1.11.4
python-dotenv==1.0.0
